pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "pillow>=12.1.1",
    "pytesseract>=0.3.13",
    "python-whois>=0.9.6",
    "requests>=2.33.0",
    "uvicorn[standard]>=0.42.0",
]
